поменял структуру проекта(будет еще меняться)+ написал конфиг, коннект к базе, выгрузку данных из базы и логгер
