+++ config_example.yml
@@ -1,0 +1,9 @@
+service:
+  logLevel: "DEBUG"
+  interval: 60 # интервал загрузки данных
+db:
+  user: "user"
+  password: "password"
+  host: "localhost"
+  port: "5432"
+  name: "db_name"
