+++ file.log
@@ -1,0 +1,14 @@
+
+2024-10-23T15:38:08.007703+0300 INFO Set logger and config successful
+2024-10-23T15:38:08.077716+0300 INFO Connected to database successfully
+2024-10-23T15:38:08.421708+0300 INFO read data successfully
+2024-10-23T15:38:08.421883+0300 INFO get data from db successful
+2024-10-23T15:38:08.432856+0300 DEBUG Top of DataFrame:
+   Суммарная нагрузка по весам, т/ч  ...  Нагрузка на 2 половину с циркуляци
+0                        678.776135  ...                          328.538409
+1                        678.509057  ...                          316.242558
+2                        681.439898  ...                          310.129909
+3                        678.913791  ...                          300.840116
+4                        681.814563  ...                          301.191280
+
+[5 rows x 15 columns]
